+++ demo/hexagonal/components/Hexagon.js
@@ -1,4 +1,4 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
 
 const Hexagon = ({ alive, onClick, x, y }) => (
   <polygon
@@ -8,13 +8,6 @@ const Hexagon = ({ alive, onClick, x, y }) => (
     points='38.9068174 -5 71.8136348 13.9987599 71.8136348 51.9962797 38.9068174 70.9950396 6 51.9962797 6 13.9987599'
   />
 )
-
-Hexagon.propTypes = {
-  alive: PropTypes.bool.isRequired,
-  onClick: PropTypes.func.isRequired,
-  x: PropTypes.number.isRequired,
-  y: PropTypes.number.isRequired
-}
 
 Hexagon.defaultProps = {
   alive: true,